pyproject.toml
[project]
name = "fastapi-at1"
version = "0.1.0"
description = "A first introduction to fullstack applications and database handling."
authors = [
    {name = "Icecreambobcat",email = "icekingrex123@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12.0, <3.13"
dependencies = [
    "fastapi (>=0.120.4,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "openai (>=2.9.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
